+++ CHANGELOG.md
@@ -48,7 +48,8 @@ Please mark all change in change log and use the issue from GitHub
 -   \#2185 Change id to string format in http module
 -   \#2186 Update endpoints in http module
 -   \#2190 Fix memory usage is twice of index size when using GPU searching
--   \#2252 Upgrade mishards to v0.9.0
+-   \#2248 Use hostname and port as instance label of metrics
+-   \#2252 Upgrade mishards APIs and requirements
 -   \#2256 k-means clustering algorithm use only Euclidean distance metric
 
 ## Task